+++ src/index.ts
@@ -2,9 +2,10 @@ export { useForm } from "./useForms";
 export { useFields } from "./useFields";
 export { useErrors } from "./useErrors";
 export {
-    useHandlers, useSettersAsEventHandler, useValidatorWithValues, useSettersAsRefEventHandler,
+    useHandlers, useSettersAsEventHandler, useSettersAsRefEventHandler,
 } from "./useHandlers";
 export { useSubmission } from "./useSubmission";
 export { useTouch } from "./useTouch";
 export { useValidation } from "./useValidation";
 export { useConstraints } from "./useConstraints";
+export { useNormalizers, normalizeNestedObjects } from "./useNormalizers";