+++ src/useFields.ts
@@ -1,13 +1,35 @@
-import { useCallback } from "react";
+import { useCallback, useMemo } from "react";
 import {
-    useGenericValues, UseResetableValuesHook, Values, MutableValues,
+    useGenericValues, UseResetableValuesHook,
 } from "./useGenericValues";
 
-export type userSuppliedValue = string | string[] | boolean | number | undefined | null;
-export type Fields = Values<userSuppliedValue>;
-export type MutableFields = MutableValues<userSuppliedValue>;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type FieldValue = string | number | boolean | undefined | MutableFields | any[];
 
-function getResetValue(currentValue: userSuppliedValue): userSuppliedValue {
+export type MutableFields = Partial<{
+    [key: string]: FieldValue;
+}>;
+export type Fields = Readonly<MutableFields>;
+
+interface SetField {
+    (name: string, value: FieldValue, eventTarget?: EventTarget | null): void;
+}
+export interface UseFieldsHook extends Omit<UseResetableValuesHook<FieldValue>, "setValue"> {
+    readonly setValue: SetField;
+}
+
+export type NormalizeSetValue = Readonly<{
+    name: string;
+    value: FieldValue;
+    currentValues: Fields;
+    eventTarget?: EventTarget | null;
+    normalizeName: (normalizedName: string) => void;
+}>;
+export interface NormalizerHandler {
+    (valuesUpdate: NormalizeSetValue): FieldValue;
+}
+
+function getResetValue(currentValue: FieldValue): FieldValue {
     switch (typeof currentValue) {
     case "number":
         return 0;
@@ -22,9 +44,39 @@ function getResetValue(currentValue: userSuppliedValue): userSuppliedValue {
 }
 
 // eslint-disable-next-line import/prefer-default-export
-export function useFields(initialValues?: Fields): UseResetableValuesHook<userSuppliedValue> {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { resetValues: _, setValues, ...resetableValues } = useGenericValues(initialValues);
+export function useFields(
+    initialValues?: Fields,
+    normalizer?: NormalizerHandler,
+): UseFieldsHook {
+    const {
+        setValues,
+        setValue: setGenericValue,
+        ...resetableValues
+    } = useGenericValues(initialValues);
+    const setValue = useMemo(() => {
+        if (typeof normalizer === "function") {
+            return (name: string, value: FieldValue, eventTarget?: EventTarget | null): void => {
+                setValues((currentValues: Fields): Fields => {
+                    let normalizedName = name;
+                    const normalizedValue = normalizer({
+                        name,
+                        value,
+                        currentValues,
+                        eventTarget,
+                        normalizeName(newName: string) {
+                            normalizedName = newName;
+                        },
+                    });
+                    return {
+                        ...currentValues,
+                        [normalizedName]: normalizedValue,
+                    };
+                });
+            };
+        }
+        return setGenericValue;
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, []);
     const resetValues = useCallback((): void => {
         setValues((currentState): Fields => {
             const nonNullInitialValues: MutableFields = { ...initialValues } || {};
@@ -38,5 +90,7 @@ export function useFields(initialValues?: Fields): UseResetableValuesHook<userSu
         });
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
-    return { ...resetableValues, setValues, resetValues };
+    return {
+        ...resetableValues, setValues, resetValues, setValue,
+    };
 }