+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* [`useNormalizers`] added a hook (`useNormalizers`) for normalizing input values
+* [`normalizeNestedObjects`] added a `useNormalizer` compatible function that handles nested objects and arrays
+
+### Changed
+* [`useSettersAsEventHandler`] Changed the third argument that is invoked with each handler from an Event to an EventTarget
+
+## [0.0.8] - 2019-09-1
 ### Fixed
 * [`useForm`, `useFields`] Resetting fields now sets the value to empty string if no default value